requirements.txt
Flask==3.0.3
PyMuPDF==1.24.10
pdfplumber==0.11.4
Pillow==10.4.0
python-docx==1.1.2
supabase==2.7.4
python-dotenv==1.0.1
requests
deep-translator==1.11.4
gTTS==2.5.4